Corrects name for Section 8 program in comments
Not sure about what it should be on the page itself - it's called
"Section 8 Project-Based Voucher program", but that doesn't say
anything about housing, so that's kind of useless. On the other
hand, calling it "Section 8 Housing" isn't accurate and would make
it harder to look up afterwards if someone wanted to.