MUST FIX bugs in housing.js calculations. Provides necessary
fields in expenses form that must be used in housing calculations
(and some SNAP ones too). Based on google spreadsheet mockup of
combo form at http://bit.ly/2wQNcVY. Bugs found in housing.js -
line 253 needs max instead of min and 'current' values MUST be
defaulted to 'previous' values.

When to create default values? When 'previous' values are created,
or in every single place 'current' values are used? If the former,
it will be prone to error if code is changed by hand for testing
since their components might not be run. Maybe find a different
centralized place.