Strengthens current expenses, maybe fixes snap calcs, `.isOver12()`
SNAP calculations for having dependents was wrong, though it's not
yet clear if they're fixed at this point. Testing further down is
needed, or possibly more abstraction.

Use as few direct member tests as possible outside of
'getMembers.js'. If members are changed in future, there will be
fewer issues with maintanance.