Adjusts styling - will not look like column spacing till radio
buttons are put on the left. If we style using width, things don't
adjust properly. Making something a percent width relies on the
parent being a certain width. Setting a max-width relies on the
label needing a certain width. By putting the radio buttons on the
left, we can assure the same width for the left-hand column. My bad
for forgetting that reasoning completely when we were talking about
putting the radio buttons on the left.

If we go back to the old style, the toggles on the expenses page
get squashed.