Fix blocked navigation when forms untouched
During the recent refactor of the various prompts, the React Router
Prompt was no longer passed the when prop, and so was not aware when the
prompt should and should not be shown. This caused navigation attempts
handled by React Router to always be blocked.

The solution was to once again provide the when prop; this value is
called isBlocking in VisitPage.